Renamed the standard project.
