+++ EnumStringValueTests/EnumStringValueTests.csproj
@@ -55,6 +55,12 @@
   <ItemGroup>
     <None Include="packages.config" />
   </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\EnumStringValues\EnumStringValues.csproj">
+      <Project>{eb9b3169-1860-4a12-b342-677accffb64f}</Project>
+      <Name>EnumStringValues</Name>
+    </ProjectReference>
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
