+++ EnumStringValues/EnumStringValues.csproj
@@ -2,6 +2,7 @@
 
   <PropertyGroup>
     <TargetFrameworks>netstandard2.0;net35;net40;net451;net461</TargetFrameworks>
+    <AssemblyName>EnumStringValues</AssemblyName>
   </PropertyGroup>
 
 </Project>
